refactor: move RolloutLimits and RolloutSession out of interception
They are rollout-domain (per-rollout budget; the per-rollout unit of ctx,
trace, stops, limits, user, error) — the interception server consumes a
session, it doesn't own the concept. New verifiers.v1.session module; the
interception package now holds only connectivity: contract, server, pools,
tunnels, configs.
