+++ verifiers/v1/interception/server.py
@@ -24,16 +24,14 @@ import contextlib
 import json
 import logging
 import secrets
-from collections.abc import AsyncIterator, Awaitable, Callable
+from collections.abc import AsyncIterator
 from contextlib import asynccontextmanager
-from dataclasses import dataclass, field
-from typing import TYPE_CHECKING, Literal
+from typing import Literal
 
 from aiohttp import web
 from pydantic import TypeAdapter, ValidationError
 from pydantic_core import PydanticSerializationError, from_json, to_json
 
-from verifiers.v1.clients import ModelContext
 from verifiers.v1.dialects import DIALECTS, Dialect
 from verifiers.v1.dialects.base import is_sse_done_event
 from verifiers.v1 import graph
@@ -45,11 +43,8 @@ from verifiers.v1.errors import (
 )
 from verifiers.v1.interception.base import BaseInterceptionConfig, Interception, Slot
 from verifiers.v1.interception.tunnel import PrimeTunnelConfig, Tunnel, TunnelConfig
-from verifiers.v1.trace import Trace
+from verifiers.v1.session import RolloutSession
 from verifiers.v1.types import Messages, Tool
-
-if TYPE_CHECKING:
-    from verifiers.v1.mcp import Respond
 
 logger = logging.getLogger(__name__)
 
@@ -84,82 +79,6 @@ async def _queue_chunks(
     finally:
         await queue.put(None)
         ready.set()
-
-
-@dataclass(frozen=True)
-class RolloutLimits:
-    """Per-rollout framework limits (None = no cap), checked before each turn is served.
-    The first limit reached refuses the turn — halting any harness, the same mechanism as
-    a @stop — and becomes the trace's stop condition. Each caps a trace computed property:
-    `max_turns` -> num_turns, `max_input_tokens` -> num_input_tokens, `max_output_tokens` ->
-    num_output_tokens, `max_total_tokens` -> num_total_tokens. Token caps are soft by one turn:
-    they're checked between turns, so the turn that crosses a cap still completes."""
-
-    max_turns: int | None = None
-    max_input_tokens: int | None = None
-    max_output_tokens: int | None = None
-    max_total_tokens: int | None = None
-
-    def reached(self, trace: Trace) -> str | None:
-        """The name of the first limit `trace` has reached, or None if within all caps."""
-        if self.max_turns is not None and trace.num_turns >= self.max_turns:
-            return "max_turns"
-        if (
-            self.max_input_tokens is not None
-            and trace.num_input_tokens >= self.max_input_tokens
-        ):
-            return "max_input_tokens"
-        if (
-            self.max_output_tokens is not None
-            and trace.num_output_tokens >= self.max_output_tokens
-        ):
-            return "max_output_tokens"
-        if (
-            self.max_total_tokens is not None
-            and trace.num_total_tokens >= self.max_total_tokens
-        ):
-            return "max_total_tokens"
-        return None
-
-
-@dataclass
-class RolloutSession:
-    ctx: ModelContext
-    trace: Trace
-    stops: list[Callable[[Trace], Awaitable[bool]]] = field(default_factory=list)
-    limits: RolloutLimits = field(default_factory=RolloutLimits)
-    user: "Respond | None" = None
-    """A user simulator the rollout sets before the harness runs (see `verifiers.v1.mcp.user`).
-    When set, each model turn with no tool call is followed by the simulator's reply,
-    injected as a user turn, and the model is re-prompted — all within one program request,
-    transparently to the harness."""
-    opening: Messages | None = None
-    """Cached opening `respond("")` messages for a no-prompt task. Computed once and re-injected on
-    every request until the first turn lands on the trace — so a retried opening request (e.g. the
-    harness SDK retrying a transient model 502, before any turn is recorded) never calls `respond`
-    twice and advances the simulator's queue past the opening."""
-    error: "RolloutError | None" = None
-    """The latest unresolved model-call failure. The harness only sees it as an HTTP error
-    (and may swallow it, or exit non-zero), so the rollout re-raises this original error once the
-    harness returns — recording the real `ProviderError` instead of a secondary `HarnessError`.
-    Reset before each model turn, so a successful retry clears it."""
-
-    async def refused(self) -> str | None:
-        """The framework's limits (turns / token budget) and `@stop` checks, run before each
-        model call. Sets the stop condition and returns its name, else None. A refused first
-        call halts the harness (its model call errors out); Harness.run treats it as clean. A task
-        that ends a trajectory from `trace.state` does it with its own `@stop` (run here generically),
-        so the interception server holds no opinion about the state's contents."""
-        if (limit := self.limits.reached(self.trace)) is not None:
-            self.trace.stop(limit)
-            logger.debug("limit %r reached: id=%s", limit, self.trace.id)
-            return limit
-        for stop in self.stops:
-            if await stop(self.trace):
-                self.trace.stop(stop.__name__)
-                logger.debug("stop %r fired: id=%s", stop.__name__, self.trace.id)
-                return stop.__name__
-        return None
 
 
 class InterceptionServerConfig(BaseInterceptionConfig):
